move s3 credential manager into protocols/s3
