+++ src/storage/protocols/s3/credentials-manager/index.ts
