+++ src/storage/protocols/s3/credentials-manager/store-knex.ts
